feat: enhance browser translation with dynamic translation hook
- Add useDynamicTranslate hook for experimental browser translation
- Implement toggle-able translation with visual indicators (🌐/🔄)
- Apply dynamic translation to node descriptions and version changelogs
- Convert i18n hook from TypeScript to TSX for React component support
- Update dependencies to support new translation features
