+++ package.json
@@ -62,6 +62,7 @@
     "firebase": "^10.14.1",
     "flowbite": "^2.5.2",
     "flowbite-react": "0.7.5",
+    "hot-memo": "^1.1.1",
     "i18n-extract": "^0.6.7",
     "i18next": "^25.5.0",
     "i18next-browser-languagedetector": "^8.2.0",
@@ -97,6 +98,7 @@
     "sflow": "^1.24.5",
     "sharp": "^0.34.3",
     "styled-jsx": "^5.1.7",
+    "use-async": "^1.2.0",
     "yaml": "^2.8.1",
     "zod": "^3.25.76"
   },
